Add tests for entity with adapter #35
