+++ src/db/set_db_entity_str_adapter.py
@@ -8,8 +8,8 @@ from src.models.ip_address_transformation import IPv4AddressStrTransformer
 from src.models.ip_network_transformation import IPv4NetworkStrTransformer
 from src.models.uuid_transformation import UUIDStrTransformer
 
-from .base_set_db_entity import BaseSetDbEntityStrAdapter
 from .base_set_db_entity import ISetDbEntity
+from .base_set_db_entity_adapter import BaseSetDbEntityStrAdapter
 
 
 class SetDbEntityStrAdapterUUID(BaseSetDbEntityStrAdapter[UUID, UUID], ISetDbEntity[UUID, UUID]):
